feat/refactor:: implementação de funcionalidade de reset de senha com token; refatoração de pontos do projeto; adicionado dependência do flyway para versionamento e evolução do banco de dados; e adiciona testes unitários para serviços, produto e usuário.

+++ backend/src/main/java/com/api/ICPAEcommerce/domain/product/DetailProductDTO.java
@@ -1,12 +1,14 @@
 package com.api.ICPAEcommerce.domain.product;
 
+import java.math.BigDecimal;
+
 public record DetailProductDTO (
         Long id,
         String code,
         String name,
         String description,
-        double price,
-        int quantity,
+        BigDecimal price,
+        Integer quantity,
         EnumProductCategory enumProductCategory
 
 ){
