+++ backend/src/main/java/com/api/ICPAEcommerce/domain/product/ListProductDTO.java
@@ -1,13 +1,15 @@
 package com.api.ICPAEcommerce.domain.product;
 
 
+import java.math.BigDecimal;
+
 public record ListProductDTO(
         Long id,
         String code,
         String name,
         String description,
-        double price,
-        int quantity,
+        BigDecimal price,
+        Integer quantity,
         EnumProductCategory enumProductCategory
 ) {
     public ListProductDTO(Product product) {
