+++ backend/src/main/java/com/api/ICPAEcommerce/domain/product/ProductDTO.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ProductDTO (
 
     @NotBlank
@@ -17,10 +19,10 @@ public record ProductDTO (
     String description,
 
     @DecimalMin("0.01")
-    double price,
+    BigDecimal price,
 
     @Min(1)
-    int quantity,
+    Integer quantity,
 
     @NotNull
     EnumProductCategory enumProductCategory
